==> 1.csv_creation.ipynb <==

###  In this notebook we will clean all the data and put it at on place in sql file. 
    Structure of sql file.
        -- jokes
        -- ratings
        
    We have jokes html file and ratings csv file. We will extract the joke and put it into sql data base as jokes table
    We will put all the ratings into ratings table


    Cleaning HTML file to extract joke data!

import re
import os
import codecs
import pandas as pd
import html2text
import numpy as np

file = 'init1.html'
data = codecs.open('data/raw/jokes/' + file, 'r', encoding="cp1252")
joke_html = data.read()

# Extracting joke
joke = html2text.html2text(joke_html)
# Extracting joke_id
joke_id = int(file.split('init')[1].split('.html')[0])

print(joke_id)
print(joke)

# creating a method that input the location of html file and output joke 
def joke_extractor(file):
    data = codecs.open(file, 'r', encoding="cp1252")
    joke_html = data.read()
    joke = html2text.html2text(joke_html)
    joke_id = int(file.split('init')[1].split('.html')[0])
    return (joke_id, joke)

files = os.listdir('data/raw/jokes')
jokes_list = [joke_extractor('data/raw/jokes/'+ file) for file in files if file.endswith('.html') ]
df = pd.DataFrame(jokes_list, columns=['joke_id', 'joke'])
df.head()

df.to_csv('data/jokes.csv', index = None)

### Saving ratings data into sql table !!
-------------------------------------------------

dataframe = pd.DataFrame()

df1 = pd.read_excel('data/raw/jester-data-1.xls', header = None) 
#df2 = pd.read_excel('data/raw/jester-data-2.xls', header = None) 
#df3 = pd.read_excel('data/raw/jester-data-3.xls', header = None) 
dataframe =  pd.concat([df1], axis=0)

dataframe.columns

    Since ratings have no user name or any user identification tag, we we give them user id 

user_id = np.linspace(1, len(dataframe), len(dataframe))
dataframe.insert(0, 'user_id', user_id)
dataframe = dataframe.rename({0: 'number_of_jokes_rated'}, axis='columns')
for i in range(1, 100):
    dataframe = dataframe.rename({i: 'joke_'+str(i)}, axis='columns')
dataframe = dataframe.astype({"user_id": int})
dataframe.head()

#### Also saving a copy of data into csv format !

dataframe.to_csv('data/jester_jokes_rating.csv', index = None)

-----------------------------------------------
                                                        END

-----------------------------------
