==> 2.Data pre-preprocessing.ipynb <==
#### Data Preprocessing

    In this notebook we will be doing data pre-processing. Before applying user user collaborative filtering, it's a recommended step to normalize the data, In this notebook we will be normalizing all the records and save it into sql database. 
    
    Normalization step have high complexity and takes lot of time to compute! So we will save the data into sql database


#### 1. Necessary imports

import numpy as np
import pandas as pd

ratings_df = pd.read_csv('./data/jester_jokes_rating.csv')
ratings_df.head()

#### Step 1 : Normalization of rating around zero mean

Adjusting each user's rating vector by subtracting user's mean rating from each of there ratings(This corrects for the fact that some thinks that 5 star is anything worth seeing and other thinks 3 star is good enough)

Normalizing the ratings around 0 means that the ratings get a positive or negative feel
to them.

# We will be using small set of data to monitor each step and later create a method 
# r_df is subset of rating_df ! here we will only take 10 users and 8 jokes
r_df = ratings_df.iloc[0:10, 0:10]
r_df.head()

# Since 99 represents null value, We will replace it with blank for joke_1 column
r_df['joke_1'] = r_df['joke_1'].replace([99],np.nan)
r_df

r_df.iloc[1, 2:] - np.mean(r_df.iloc[1, 2:])

# Selecting all joke columns
joke_columns = r_df.columns[2:] # joke column starts from second column
joke_columns

#### Creating the method for normalization as learnt in the tutorial above!

# replacing 99 with nan value
def replace_99(ratings):
    '''
    Replace 99 with nan for user's rating value
    step 1: select joke columns
    step 2: replace 99 with np.nan for all columns
    
    '''
    joke_ids = ratings.columns[2:]    
    for joke_id in joke_ids: 
        ratings[joke_id] = ratings[joke_id].replace([99],np.nan)
    return ratings

# Now creating a function that takes all the raw data as input and complete step 1 
def normalization(ratings):
    ''' 
    Returns mean subtracted user's rating value
    step 1: subtract user's rating by their mean value for each row 
    '''
    total_users = ratings.shape[0]
    for i in range(total_users):
        ratings.iloc[i, 2:] = ratings.iloc[i, 2:] - np.mean(ratings.iloc[i, 2:])
    return ratings

# replacing nan with 0
def replace_0(ratings):
    '''
    Replace nan with 0 for user's rating value
    step 1: select joke columns
    step 2: replace np.nan with 0 for all columns
    
    '''
    joke_ids = ratings.columns[2:]    
    for joke_id in joke_ids: 
        ratings[joke_id] = ratings[joke_id].replace([np.nan],0)
    return ratings

## Here Spark will be better choce that
temp1 = replace_99(r_df)
temp2 = normalization(temp1)
temp3 = replace_0(temp2)
temp3.head()

#### Now transformation of rating data for into normalized ratings for all the ratings! 

# we will normalize and save into normalized_ratings dataframe!
normalized_ratings = replace_99(ratings_df)
normalized_ratings = normalization(normalized_ratings)
normalized_ratings = replace_0(normalized_ratings)
normalized_ratings.head()

normalized_ratings.to_csv('data/normalized_ratings.csv', index = None)

## Finding similarity: 
We will be using pearson correlation as similarity measure !

ratings_of_active_user = temp3.iloc[6, 2:]
ratings_of_other_user = temp3.iloc[7, 2:]

# note probably use list and assert that x and y are list!
# since we are using normalized data, we won't subtract by mean value again
def weight_factor(x, y):
    ''' 
    Weight factor implies relationship between user x and user y
    Also know as similarity between user x and user y
    We are using Pearson correlation coefficient here. 
    '''    
    t1, t2, t3 = 0, 0, 0 
    for i, j in zip(x, y):
        t1+=i*j
        t2+=i*i
        t3+=j*j
    return t1/(np.sqrt(t2) * np.sqrt(t3))

    Testing out similarity 

# now our active user is suppose user_id 1!
# we want to find similarity with other users!

# x represent active user
# y represents other user
x =  temp3.iloc[0, 2:]
similarity = [weight_factor(x, temp3.iloc[i, 2:]) for i in range(temp3.shape[0])]
similarity

----------------------------------------------------------------
                                        END
                                        
-----------------------------------------------------------------
