==> 3.UUCF.ipynb <==
### Collaborative Filtering
-----------------------------------------------------------------------
----------------------------------------------------------------------

    Collaborative Filtering techniques make recommendations for a user based on ratings and preferences data of many users. The main underlying idea is that if two users have both rated 2 items similarly, then the items that one user has liked that the other user has not yet 

##### In this notebook we will apply user user collaborative filtering technique!

#### User User collaborative filtering !!
------------------------------------------

    This approach relies on the idea that users who have similar rating behaviours so far, share the same tastes and will likely exhibit similar rating behaviours going forward. The algorithm first computes the similarity between users by considering ratings both users have in common

From notebook 2, we are taking similarity funcition. We will use this function to find similarity between active user  and other user using normalized data

def weight_factor(x, y):
    ''' 
    Weight factor implies relationship between user x and user y
    Also know as similarity between user x and user y
    We are using Pearson correlation coefficient here. 
    '''    
    t1, t2, t3 = 0, 0, 0 
    for i, j in zip(x, y):
        t1+=i*j
        t2+=i*i
        t3+=j*j
    return t1/(np.sqrt(t2) * np.sqrt(t3))

#### Selecting normalized data from database!

import numpy as np
import pandas as pd

normalized_ratings_df = pd.read_csv('./data/normalized_ratings.csv')

ratings_df = pd.read_csv('./data/jester_jokes_rating.csv')

normalized_ratings_df.head(2)

#### Separating normalized ratings dataframe!

We will be separating normalized ratings dataframe into 2 parts, 

    1> Complete ratings: Those users who have rated all 100 jokes
    2> Sparse ratings: Those users who haven't rated all 100 jokes
    
For ease in computation, we will only select active user from sparse ratings, and other users from compete rating groups

# We will be using users who have rated all the 100 jokes as other users.
complete_ratings = normalized_ratings_df[normalized_ratings_df['number_of_jokes_rated'] == 100]
print('total user count who have rated all the jokes: ', len(complete_ratings))
# We will be randomly using one out of these users as active user and use it to find 
# similarity with complete_ratings dataset. 
sparse_ratings = normalized_ratings_df[normalized_ratings_df['number_of_jokes_rated'] != 100]
print('total user count who have not rated all the jokes: ', len(sparse_ratings))

    Selecting one of the user from sparse_ratings matrix for whom the recommendation will be given. We will call him our active user and generate recommendation for him!

# selecting a random user say 1000th user in sparse_ratings list
n = 1000
active_user_id = sparse_ratings.iloc[n, 0]
print("Let's selct a random user with user id {} as active user for\
 which we will recommend the joke".format(str(active_user_id)))

print('ratings given by active user {} for 100 jokes'.format(str(active_user_id)))
active_user = sparse_ratings[sparse_ratings['user_id'] == active_user_id]
active_user_rating = active_user.iloc[:, 2:]
active_user_rating

##### Finding similarity between active user and complete rated user!

# saving active user ratings into 1 d list
active_user_rating_list = active_user_rating.values.ravel()
# finding similarity between active user and all its neighbours among complete rating users
similarity = np.array([(complete_ratings.iloc[i, 0],\
             weight_factor(active_user_rating_list, complete_ratings.iloc[i, 2:]))\
             for i in range(complete_ratings.shape[0])])

#### Sorting the neighbours using similarity 

ind = np.argsort( similarity[:,1] )
similarity = similarity[ind]
similarity

### Now we have similarity matrix, our next task is to select neighbourhood. 
##### There are few methods to select neighbours:
    1. Use all neighbours in samll dataset
    2. Threshold similarity or distance:
        eg. all users with similarity of 0.1 or above to user for whom we are generating the recommendations will be considered
    3. Random neighbour:
        Can be useful for very large dataset in conjucture with other technique. for eg. randomly select 10,000 neighbour and then threshold them or pick top n
    4. Top n neighbour by similarity or distance
        
### How many neighbours?
##### Between 25 to 100 is ofen used.
        
#### Our method !
    We will be use threshold method followed by random n samples, (where n can be any random number between 25 - 100) but for simplicity we will take n as 30. 
    For threshold we will select other users with similarity > 0.1
    

neighbours = similarity[similarity[:,1] > 0.1]
print('We have {} potential neighbours! Now we will be randomly selecting 30 samples out of them'.format(len(neighbours)))

#### Randomly selecting 30 neighbours

# by replace = False, ensuring that no duplicate neighbour is selected !!
index_30_neighbour = np.random.choice(range(len(neighbours)), 30, replace=False)
selected_neighbours = neighbours[index_30_neighbour]
selected_neighbours

    Once we have selected 30 closest neighbours, our task is to create scoring for the jokes !! We will crate scoring for only those jokes which isn't rated yet by the active user yet. 

recommendation_columns = [column for column in active_user_rating.columns if active_user_rating[column].values[0] == 0]

##### Mean value of ratings of active user!!

    The mean value of rating is used to create so as to compensate for the offset created when data was normalized.
    By adding mean value to the predicted score, it will give the probable score the user will give to a joke!

active_user_raw_ratings = ratings_df[ratings_df['user_id'] == active_user_id].iloc[:, 2:]
active_user_mean_rating = np.mean(active_user_raw_ratings.drop(recommendation_columns, axis = 1).values)
active_user_mean_rating

# Selecting neighbours user_id
neighbour_user_id = selected_neighbours[:, 0]


# selectig neighbour user similarity
neighbour_user_similarity = selected_neighbours[:, 1]

# viewing all neighbours user id and similarity
print('neighbours user id: ', neighbour_user_id, '\n\n')
print('neighbours user similarity: ', neighbour_user_similarity)

#### Selecting all the data of neighbour!

neighbours_df = complete_ratings[complete_ratings['user_id'].isin(neighbour_user_id)]
len(neighbours_df)

# selecting only recommendation columns
print('We will be suggesting one out of {} jokes to the active user \n\n'.format(len(recommendation_columns)))

neighbours_df = neighbours_df[recommendation_columns]
neighbours_df.head()

#### checking score for item 0 of the recommendation_columns

item_id = recommendation_columns[0]
print( 'item 0 for which we calculate the score is', item_id)

def score_user_item(item_id, neighbours_df,neighbour_user_similarity, active_user_mean_rating ):
    item_rating = neighbours_df[item_id]
    t1, t2 = 0, 0
    for similarity, norm_rating in zip(neighbour_user_similarity, item_rating):
        t1+= norm_rating * similarity
        t2+= similarity
    score = (t1 + active_user_mean_rating)/t2
    return score

# checking score for a particular joke!
score_user_item(item_id, neighbours_df,neighbour_user_similarity, active_user_mean_rating )
        

#### suggesting use the joke which has the highesht score among all

# Computing user item score !
top_score = -np.inf
joke_to_suggest = ''

for column in neighbours_df.columns:
    score =score_user_item(column, neighbours_df,neighbour_user_similarity, active_user_mean_rating)
    if score > top_score:
        top_score = score
        joke_to_suggest = column
print('highest score is', top_score)
print('The highest score obtained by the joke among all the unseen jokes is', joke_to_suggest, )
